+++ portal/templates/layouts/navbar.html
@@ -4,6 +4,7 @@
     <span>Hello, {{ g.user[2] }}!</span>
     <li><a href="{{ url_for('courses.course_page') }}">Courses</a></li>
     <li><a href="{{ url_for('dash') }}">Dashboard</a></li>
+    <li><a href="{{ url_for('sessions.sessions') }}">Sessions</a></li>
     <li><a href="{{ url_for('auth.log_out') }}">Log Out</a></li>
   </ul>
 </nav>
